+++ noodles-cram/examples/cram_write_async.rs
@@ -66,7 +66,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let header = build_header(&reference_sequences)?;
 
     let repository = fasta::Repository::new(reference_sequences);
-    let mut writer = cram::r#async::writer::Builder::default()
+    let mut writer = cram::r#async::io::writer::Builder::default()
         .set_reference_sequence_repository(repository)
         .build_with_writer(io::stdout());