+++ noodles-cram/src/async.rs
@@ -1,6 +1,3 @@
 //! Async CRAM.
 
 pub mod io;
-pub mod writer;
-
-pub use self::writer::Writer;