+++ noodles-cram/src/async/io.rs
@@ -1,5 +1,6 @@
 //! Async CRAM I/O.
 
 mod reader;
+pub mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};